+++ src/las.h
@@ -125,7 +125,7 @@ struct node_chunk;
 void free_all_nodes(void);
 void debug_free_node(node *n);
 node *debug_check_node_hash(node *n);
-node *debug_mknode(short token,node *a,node *b);
+node *debug_mknode(int token,node *a,node *b);
 node *debug_mkstrnode(struct pike_string *str);
 node *debug_mkintnode(INT_TYPE nr);
 node *debug_mknewintnode(INT_TYPE nr);
